Split shared logic into dedicated files
